Broke out more files, builds on both OSX and Linux
