+++ Sources/vapor/array_ext.swift
@@ -1,0 +1,10 @@
+
+extension Array where Element: Equatable {
+    mutating func remove(_ element: Element) {
+        self = self.filter { $0 != element }
+    }
+
+    mutating func remove(matching: (Element) -> Bool) {
+        self = self.filter { !matching($0) }
+    }
+}
